fixing build process
