+++ package.json
@@ -24,39 +24,39 @@
   },
   "homepage": "https://github.com/kstf/plump-store-rest#readme",
   "devDependencies": {
-    "@types/bluebird": "^3.0.37",
+    "@types/bluebird": "^3.5.2",
     "@types/chai": "^3.4.35",
     "@types/chai-as-promised": "0.0.30",
-    "@types/core-js": "^0.9.38",
+    "@types/core-js": "^0.9.41",
     "@types/hapi": "^16.0.0",
     "@types/joi": "^10.0.1",
     "@types/mocha": "^2.2.40",
-    "@types/node": "^7.0.10",
-    "boom": "^4.2.0",
+    "@types/node": "^7.0.12",
+    "boom": "^4.3.1",
     "chai": "^3.5.0",
-    "chai-as-promised": "^5.3.0",
+    "chai-as-promised": "^6.0.0",
     "del": "^2.2.2",
     "gulp": "gulpjs/gulp.git#4.0",
-    "gulp-sourcemaps": "^1.6.0",
+    "gulp-sourcemaps": "^2.5.1",
     "gulp-typescript": "^3.1.6",
-    "hapi": "^16.1.0",
-    "joi": "^10.2.2",
-    "mocha": "^3.0.2",
-    "pg": "^6.1.4",
+    "hapi": "^16.1.1",
+    "joi": "^10.4.1",
+    "mocha": "^3.2.0",
+    "pg": "^6.1.5",
     "plump-store-postgres": "kstf/plump-store-postgres",
     "plump-strut": "kstf/plump-strut",
     "ts-node": "^3.0.2",
-    "tslint": "^4.5.1",
+    "tslint": "^5.0.0",
     "tslint-eslint-rules": "^3.5.1",
-    "typescript": "^2.2.1"
+    "typescript": "^2.2.2"
   },
   "peerDependencies": {
     "plump": "kstf/plump"
   },
   "dependencies": {
-    "axios": "^0.15.3",
-    "bluebird": "^3.4.6",
+    "axios": "^0.16.0",
+    "bluebird": "^3.5.0",
     "merge-options": "0.0.64",
-    "rxjs": "^5.0.0-beta.12"
+    "rxjs": "^5.3.0"
   }
 }
