+++ tsconfig.json
@@ -1,17 +1,16 @@
 {
     "compilerOptions": {
-      "allowJs": true,
       "outDir": "./dist",
       "target": "es5",
-      "lib": ["es2015", "dom"],
-      "types": ["node"],
+      "lib": ["es6", "dom"],
+      "types": ["node", "mocha"],
       "module": "commonjs",
       "moduleResolution": "node",
       "isolatedModules": false,
       "jsx": "react",
       "experimentalDecorators": true,
       "emitDecoratorMetadata": true,
-      "declaration": false,
+      "declaration": true,
       "noImplicitUseStrict": false,
       "noEmitOnError": true,
       "removeComments": true,
@@ -20,7 +19,7 @@
       "suppressImplicitAnyIndexErrors": true
     },
     "include": [
-      "./src/**/*", "./test/**/*"
+      "./src/**/*.ts", "./test/**/*.ts"
     ],
     "exclude": [
       "node_modules"
